Add support for setting mobile API keys dynamically
This adds an overloaded getService(int, String) method that takes the mode
and the mobile API key. This enables an alternative way of specifyin the
API key for applications where it's not known at compile time.

It's still possible to use the previous getService(int)
method and set the API key through the manifest meta-data. If the API key is
specified both in the manifest and dynamically, the dynamically set value is used.

To avoid duplicate code in three places to retrieve the API key in one of two ways,
a getMobileKey() helper method was also extracted.